% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstructure.R
\name{run_structure}
\alias{run_structure}
\title{Run STRUCTURE analysis}
\usage{
run_structure(
  file,
  k.range = 1:5,
  num.k.rep = 3,
  burnin = 1000,
  numreps = 1000,
  noadmix = FALSE,
  phased = FALSE,
  ploidy = 2,
  linkage = FALSE
)
}
\description{
Adapted from the dartR package which was based on strataG. Build for docker deployment.
}
